Update cses missing coing sum queries.cpp

+++ Persistent Segment Tree/cses missing coing sum queries.cpp
@@ -1,14 +1,14 @@
 #include <bits/stdc++.h>
 using namespace std;
- 
+
 using ll = long long;
- 
+
 struct Node {
     ll sum;
     Node *l, *r;
- 
+
     Node(ll s = 0) : sum(s), l(nullptr), r(nullptr) {}
- 
+
     Node(Node* L, Node* R) {
         l = L;
         r = R;
@@ -17,114 +17,113 @@ struct Node {
         if (R) sum += R->sum;
     }
 };
- 
-Node* build(int tl, int tr) {
+
+Node* build(ll tl, ll tr) {
     if (tl == tr) return new Node();
- 
-    int tm = (tl + tr) >> 1;
- 
+
+    ll tm = (tl + tr) >> 1;
+
     Node* left = build(tl, tm);
     Node* right = build(tm + 1, tr);
- 
+
     return new Node(left, right);
 }
- 
-Node* update(Node* node, int tl, int tr, int pos, ll val) {
+
+Node* update(Node* node, ll tl, ll tr, ll pos, ll val) {
     if (tl == tr) {
         Node* cur = new Node();
         cur->sum = node->sum + val;
         return cur;
     }
- 
-    int tm = (tl + tr) >> 1;
- 
+
+    ll tm = (tl + tr) >> 1;
+
     if (pos <= tm) {
         return new Node(
             update(node->l, tl, tm, pos, val),
             node->r
         );
     }
- 
+
     return new Node(
         node->l,
         update(node->r, tm + 1, tr, pos, val)
     );
 }
- 
-ll query(Node* node, int tl, int tr, int l, int r) {
+
+ll query(Node* node, ll tl, ll tr, ll l, ll r) {
     if (!node || l > tr || r < tl) return 0;
- 
+
     if (l <= tl && tr <= r)
         return node->sum;
- 
-    int tm = (tl + tr) >> 1;
- 
+
+    ll tm = (tl + tr) >> 1;
+
     return query(node->l, tl, tm, l, r)
          + query(node->r, tm + 1, tr, l, r);
 }
- 
+
 int main() {
     ios::sync_with_stdio(false);
     cin.tie(nullptr);
- 
-    int n, q;
+
+    ll n, q;
     cin >> n >> q;
- 
+
     vector<ll> a(n + 1);
- 
     vector<ll> vals;
- 
-    for (int i = 1; i <= n; i++) {
+
+    for (ll i = 1; i <= n; i++) {
         cin >> a[i];
         vals.push_back(a[i]);
     }
- 
+
     sort(vals.begin(), vals.end());
     vals.erase(unique(vals.begin(), vals.end()), vals.end());
- 
-    int m = vals.size();
- 
+
+    ll m = vals.size();
+
     vector<Node*> root(n + 1);
- 
+
     root[0] = build(1, m);
- 
-    for (int i = 1; i <= n; i++) {
-        int pos =
+
+    for (ll i = 1; i <= n; i++) {
+        ll pos =
             lower_bound(vals.begin(), vals.end(), a[i])
             - vals.begin() + 1;
- 
+
         root[i] = update(root[i - 1], 1, m, pos, a[i]);
     }
- 
+
     while (q--) {
- 
-        int l, r;
+
+        ll l, r;
         cin >> l >> r;
- 
+
         ll reach = 1;
- 
+
         while (true) {
- 
-            int pos =
+
+            ll pos =
                 upper_bound(vals.begin(), vals.end(), reach)
                 - vals.begin();
- 
+
             ll s = 0;
- 
+
             if (pos > 0) {
                 s =
                     query(root[r], 1, m, 1, pos)
                   - query(root[l - 1], 1, m, 1, pos);
             }
- 
+
             if (s < reach) {
                 cout << reach << '\n';
                 break;
             }
- 
+
             reach = s + 1;
         }
     }
- 
+
     return 0;
 }
